+++ apps/storybook/.storybook/preview.ts
@@ -2,6 +2,9 @@ import type { Preview } from '@storybook/react';
 import '../../../packages/react/src/styles/variables.css';
 import '../../../packages/react/src/styles/reset.css';
 import '../../../packages/react/src/components/NavigationBar.css';
+import '../../../packages/react/src/components/Badge.css';
+import '../../../packages/react/src/components/Button.css';
+import '../../../packages/react/src/components/HeroSection.css';
 
 const preview: Preview = {
   parameters: {